pyproject.toml
[project]
name = "instacrawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "instaloader>=4.15",
    "neo4j>=6.0.3",
    "pillow>=12.0.0",
]
